Move logging and error modules to their own directories
