+++ .changeset/angry-views-draw.md
@@ -1,0 +1,5 @@
+---
+'@bengsfort/stdlib': minor
+---
+
+BREAKING: Moves logging and error modules from top level to their own directories.
