+++ lib/errors/helpers.ts
