+++ lib/logging/internal/utils.ts
